richMenuのprops調整
