+++ src/components/molecules/RichMenuList/type.ts
@@ -3,7 +3,7 @@ import { ListItemProps } from "@components/atoms/ListItem/type";
 
 export type RichMenuListProps = {
   onClickButton: () => void;
-  groups: {
+  groups?: {
     id: string;
     title: string;
     titleIcon?: ReactElement;
