--- src/components/organisms/RichMenu/index.stories.ts
@@ -1,52 +1,0 @@
-import { MenuListTypeKey } from "@components/molecules/MenuList";
-import type { Meta, StoryObj } from "@storybook/react";
-
-import { RichMenu } from "./";
-
-const RICH_MENU_TYPES: MenuListTypeKey[] = ["account", "beyond", "connection"];
-
-export default {
-  component: RichMenu,
-  parameters: {
-    controls: { expanded: true },
-  },
-  argTypes: {
-    absolute: {
-      type: "boolean",
-    },
-    isOpen: {
-      type: "boolean",
-    },
-    navigationType: {
-      type: "string",
-      options: RICH_MENU_TYPES,
-      control: {
-        type: "select",
-      },
-    },
-    onOpenChange: {
-      type: "function",
-    },
-  },
-} satisfies Meta<typeof RichMenu>;
-
-export const Default: StoryObj<typeof RichMenu> = {
-  args: {
-    isOpen: true,
-    navigationType: "beyond",
-  },
-};
-
-export const List: StoryObj<typeof RichMenu> = {
-  args: {
-    isOpen: true,
-    navigationType: "beyond",
-  },
-};
-
-export const Description: StoryObj<typeof RichMenu> = {
-  args: {
-    isOpen: true,
-    navigationType: "beyond",
-  },
-};
