+++ src/components/organisms/RichMenu/type.ts
@@ -1,4 +1,5 @@
 import type { MenuListTypeKey } from "@components/molecules/MenuList";
+import { RichMenuListProps } from "@components/molecules/RichMenuList/type";
 
 export type RichMenuProps = {
   navigationType: MenuListTypeKey;
@@ -6,4 +7,5 @@ export type RichMenuProps = {
   onOpenChange?: () => void;
   absolute?: boolean;
   anchor?: "right" | "left";
+  groups?: RichMenuListProps["groups"];
 };
